+++ NEWS.md
@@ -10,3 +10,4 @@
     * Also, now check for invalid aggregation methods in `createSlotAggList`, so it will not wait until `processSlots` is called to throw the error. 
 * ensured that `getDataForAllScens()` (and interal function `processSlots()`) work with rdf files that only include one trace of data (#40)
     * improved tests for this case
+* new function: `makeAllScenNames()`